Moved items to store directory for organization
