+++ SEBrowser/Scripts/TSX/Components/EventSearch/EventSearchNavbar.tsx
@@ -26,9 +26,9 @@ import React from 'react';
 import _ from 'lodash';
 import ReportTimeFilter from '../ReportTimeFilter';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { SelectAssetGroupList, SelectAssetList, SelectCharacteristicFilter, SelectMeterList, SelectReset, SelectStationList, SelectTimeFilter, SelectTypeFilter, SetFilterLists } from './EventSearchSlice';
-import { ResetFilters,  SetFilters } from './EventSearchSlice';
-import { AssetGroupSlice, AssetSlice, LocationSlice, MeterSlice, MagDurCurveSlice, EventTypeSlice } from '../../Store';
+import { SelectAssetGroupList, SelectAssetList, SelectCharacteristicFilter, SelectMeterList, SelectReset, SelectStationList, SelectTimeFilter, SelectTypeFilter, SetFilterLists } from '../../Store/EventSearchSlice';
+import { ResetFilters,  SetFilters } from '../../Store/EventSearchSlice';
+import { AssetGroupSlice, AssetSlice, LocationSlice, MeterSlice, MagDurCurveSlice, EventTypeSlice } from '../../Store/Store';
 import { DefaultSelects } from '@gpa-gemstone/common-pages';
 import NavbarFilterButton from '../Common/NavbarFilterButton';
 import { SystemCenter, OpenXDA } from '@gpa-gemstone/application-typings';
@@ -37,7 +37,7 @@ import { Search } from '@gpa-gemstone/react-interactive';
 import { Column } from '@gpa-gemstone/react-table';
 import { SEBrowser } from '../../Global';
 import EventSearchTypeFilters from './EventSearchTypeFilter';
-import { SelectDateTimeSetting, SelectTimeZone } from '../SettingsSlice';
+import { SelectDateTimeSetting, SelectTimeZone } from '../../Store/SettingsSlice';
 import { getMoment, getStartEndTime, readableUnit } from './TimeWindowUtils';
 
 interface IProps {
