+++ SEBrowser/Scripts/TSX/hooks.ts
@@ -22,7 +22,7 @@
 //******************************************************************************************************
 
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import { AppDispatch, RootState } from './Store';
+import { AppDispatch, RootState } from './Store/Store';
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
